Refactored hi.simulator.scenes.views
