+++ src/hi/simulator/scenes/views.py
@@ -1,6 +1,5 @@
 import json
 import logging
-import re
 
 from django.conf import settings
 from django.core.cache import cache
@@ -20,6 +19,9 @@ from .models import SimScene, SimSceneBinding, SimSceneControl, SimStateSequence
 from .player import SimPlayer
 from .recorder import SimRecorder
 from .scene_controller import SceneController
+from .control_grid import control_key, resolve_control, simulator_by_module, split_control_key
+from .sequence_io import parse_steps, safe_filename
+from .transport import build_transport_context, transport_signature
 
 logger = logging.getLogger(__name__)
 
@@ -41,279 +43,6 @@ def _get_scene( scene_id ) -> SimScene:
         return SimScene.objects.get( id = scene_id )
     except SimScene.DoesNotExist:
         raise Http404( 'Scene not found.' )
-
-
-def _simulator_by_module():
-    return {
-        simulator_data.simulator.module_key: simulator_data.simulator
-        for simulator_data in ServiceSimulatorManager().get_simulator_data_list()
-    }
-
-
-def _control_key( module_key, entity_name, sim_state_id ):
-    return f'{module_key}|{entity_name}|{sim_state_id}'
-
-
-def _split_control_key( key ):
-    # module_key (dotted) and sim_state_id (slug) are '|'-free, so take
-    # them from the ends; entity_name (the middle) may contain '|'.
-    parts = key.split( '|' )
-    return parts[0], '|'.join( parts[1:-1] ), parts[-1]
-
-
-def _resolve_control( control, simulator_by_module ):
-    """Resolve a SimSceneControl's stable key to the live (simulator,
-    sim_state); None if the entity/state isn't currently loaded."""
-    simulator = simulator_by_module.get( control.module_key )
-    if simulator is None:
-        return None
-    sim_entity = next(
-        ( e for e in simulator.sim_entities if e.name == control.entity_name ),
-        None,
-    )
-    if sim_entity is None:
-        return None
-    sim_state = next(
-        ( s for s in sim_entity.sim_state_list if s.sim_state_id == control.sim_state_id ),
-        None,
-    )
-    if sim_state is None:
-        return None
-    # The scenes grid shows the entity name in its own column, so drop a
-    # redundant leading "<entity> " from the state label (e.g. hass's
-    # "Front Door Motion" -> "Motion"). Labels without that prefix (frigate's
-    # "Detected Object") are left as-is.
-    entity_name = sim_entity.name
-    state_label = sim_state.name
-    prefix = f'{entity_name} '
-    if state_label.startswith( prefix ) and len( state_label ) > len( prefix ):
-        state_label = state_label[ len( prefix ): ]
-    return {
-        'simulator': simulator,
-        'sim_state': sim_state,
-        'entity_name': entity_name,
-        'state_label': state_label,
-    }
-
-
-def _parse_steps( raw ):
-    """Validate edited/imported sequence JSON. Accepts either a bare list of
-    steps or an export envelope ({name, steps}); returns (steps, error)."""
-    try:
-        data = json.loads( raw )
-    except ( ValueError, TypeError ):
-        return None, 'Not valid JSON.'
-    if isinstance( data, dict ) and 'steps' in data:
-        data = data[ 'steps' ]
-    if not isinstance( data, list ):
-        return None, 'Expected a JSON list of steps.'
-    cleaned = []
-    for index, step in enumerate( data ):
-        if not isinstance( step, dict ):
-            return None, f'Step {index} is not an object.'
-        if step.get( 'end' ):
-            if 't' not in step:
-                return None, f'Step {index} (end) is missing "t".'
-            try:
-                cleaned.append({ 't': float( step[ 't' ] ), 'end': True })
-            except ( TypeError, ValueError ):
-                return None, f'Step {index} (end) has a non-numeric "t".'
-            continue
-        if 'marker' in step:
-            marker = { 'marker': str( step[ 'marker' ] ) }
-            if 't' in step:
-                try:
-                    marker[ 't' ] = float( step[ 't' ] )
-                except ( TypeError, ValueError ):
-                    return None, f'Marker {index} has a non-numeric "t".'
-            cleaned.append( marker )
-            continue
-        missing = [ k for k in ( 't', 'module', 'entity', 'state' ) if k not in step ]
-        if missing:
-            return None, f'Step {index} is missing field(s): {", ".join( missing )}.'
-        try:
-            t = float( step[ 't' ] )
-        except ( TypeError, ValueError ):
-            return None, f'Step {index} has a non-numeric "t".'
-        cleaned.append({
-            't': t,
-            'module': str( step[ 'module' ] ),
-            'entity': str( step[ 'entity' ] ),
-            'state': str( step[ 'state' ] ),
-            'value': step.get( 'value' ),
-        })
-        continue
-    return cleaned, None
-
-
-def _safe_filename( name ):
-    slug = re.sub( r'[^A-Za-z0-9._-]+', '_', name.strip() ).strip( '_' )
-    return slug or 'sequence'
-
-
-# Timeline geometry, in SVG viewBox units (the SVG scales to fit via
-# preserveAspectRatio="none"; the click handler maps pixels back to time).
-TIMELINE_VBW = 1000
-TIMELINE_HEIGHT = 72
-TIMELINE_X0 = 48
-TIMELINE_X1 = 952
-TIMELINE_AXIS_Y = 44
-TIMELINE_MARKER_TOP = TIMELINE_AXIS_Y - 22    # event arrows above the axis
-TIMELINE_STEP_TIP_Y = TIMELINE_AXIS_Y + 3     # state-change arrows below the axis
-TIMELINE_STEP_BASE_Y = TIMELINE_AXIS_Y + 21
-TIMELINE_GLYPH_HALF = 8
-TIMELINE_HIT_HALF = 10                         # generous transparent hover/click column
-
-
-SERVICE_MODULE_PREFIX = 'hi.simulator.services.'
-
-
-def _short_module( module ):
-    module = module or ''
-    if module.startswith( SERVICE_MODULE_PREFIX ):
-        return module[ len( SERVICE_MODULE_PREFIX ): ]
-    return module
-
-
-def _timeline_x( t, total ):
-    if total <= 0:
-        return TIMELINE_X0
-    frac = min( 1.0, max( 0.0, t / total ))
-    return round( TIMELINE_X0 + frac * ( TIMELINE_X1 - TIMELINE_X0 ), 2 )
-
-
-def _timeline_model( sequence, player_status ):
-    """Parse a sequence's steps into a render model for the SVG timeline.
-    Markers are down-arrows above the axis; state changes are up-arrows below
-    it; the end sentinel is an end-rule at the right edge. Each glyph gets a
-    wide transparent hit column so it is easy to hover/click. Timed items sit
-    at their own t; a legacy marker without a t falls back to the preceding
-    timed step (start is the implicit 0, never materialized)."""
-    steps = sequence.steps_json or []
-    total = max( ( float( s[ 't' ] ) for s in steps if 't' in s ), default = 0.0 )
-    entries = []
-    last_t = 0.0
-    for step in steps:
-        if step.get( 'end' ):
-            et = float( step.get( 't', last_t ))
-            last_t = max( last_t, et )
-            x = _timeline_x( et, total )
-            entries.append({
-                'kind': 'end',
-                't': round( et, 3 ),
-                'x': x,
-                'hit_x': round( x - TIMELINE_HIT_HALF, 2 ),
-                'hit_y': 0,
-                'hit_h': TIMELINE_HEIGHT,
-                'title': 'End  (@{:.2f}s)'.format( et ),
-            })
-            continue
-        if 'marker' in step:
-            mt = float( step[ 't' ] ) if 't' in step else last_t
-            last_t = max( last_t, mt )
-            x = _timeline_x( mt, total )
-            entries.append({
-                'kind': 'marker',
-                't': round( mt, 3 ),
-                'x': x,
-                'hit_x': round( x - TIMELINE_HIT_HALF, 2 ),
-                'hit_y': 0,
-                'hit_h': TIMELINE_AXIS_Y,           # events: above the axis only
-                'points': '{x0},{top} {x1},{top} {x},{tip}'.format(
-                    x0 = x - TIMELINE_GLYPH_HALF, x1 = x + TIMELINE_GLYPH_HALF,
-                    top = TIMELINE_MARKER_TOP, x = x, tip = TIMELINE_AXIS_Y,
-                ),
-                'title': '⚑ {}  (@{:.2f}s)'.format( step[ 'marker' ], mt ),
-            })
-            continue
-        try:
-            t = float( step.get( 't', last_t ))
-        except ( TypeError, ValueError ):
-            t = last_t
-        last_t = max( last_t, t )
-        x = _timeline_x( t, total )
-        entries.append({
-            'kind': 'step',
-            't': round( t, 3 ),
-            'x': x,
-            'hit_x': round( x - TIMELINE_HIT_HALF, 2 ),
-            'hit_y': TIMELINE_AXIS_Y,               # transitions: below the axis only
-            'hit_h': TIMELINE_HEIGHT - TIMELINE_AXIS_Y,
-            'points': '{x},{tip} {x0},{base} {x1},{base}'.format(
-                x = x, tip = TIMELINE_STEP_TIP_Y,
-                x0 = x - TIMELINE_GLYPH_HALF, x1 = x + TIMELINE_GLYPH_HALF,
-                base = TIMELINE_STEP_BASE_Y,
-            ),
-            'title': '{:.2f}s  [{}] {}.{} = {}'.format(
-                t, _short_module( step.get( 'module' )), step.get( 'entity' ),
-                step.get( 'state' ), step.get( 'value' ),
-            ),
-        })
-        continue
-
-    active = bool(
-        player_status.get( 'loaded' )
-        and player_status.get( 'sequence_id' ) == sequence.id
-    )
-    playhead_t = player_status.get( 'playhead', 0.0 ) if active else 0.0
-    return {
-        'total': total,
-        'entries': entries,
-        'vbw': TIMELINE_VBW,
-        'height': TIMELINE_HEIGHT,
-        'x0': TIMELINE_X0,
-        'x1': TIMELINE_X1,
-        'axis_y': TIMELINE_AXIS_Y,
-        'hit_w': 2 * TIMELINE_HIT_HALF,
-        'end_y1': 2,
-        'end_y2': TIMELINE_HEIGHT - 2,
-        'playhead_y1': 6,
-        'playhead_y2': TIMELINE_HEIGHT - 6,
-        'playhead_t': round( playhead_t, 3 ),
-        'playhead_x': _timeline_x( playhead_t, total ),
-        'active': active,
-    }
-
-
-def _resolve_sequence( sequence_param, sequence_list ):
-    if sequence_param:
-        match = next(
-            ( s for s in sequence_list if str( s.id ) == str( sequence_param )), None,
-        )
-        if match is not None:
-            return match
-    return sequence_list[ 0 ] if sequence_list else None
-
-
-def _build_transport_context( scene, sequence_param ):
-    """Context for the transport pane (badge, controls, timeline) — shared by
-    the full dashboard and the live fragment-refresh endpoint."""
-    sequence_list = list( scene.state_sequences.all() )
-    selected_sequence = _resolve_sequence( sequence_param, sequence_list )
-    player_status = SimPlayer().get_status()
-    timeline = None
-    if selected_sequence is not None:
-        timeline = _timeline_model( selected_sequence, player_status )
-    return {
-        'scene': scene,
-        'sequence_list': sequence_list,
-        'selected_sequence': selected_sequence,
-        'recorder_status': SimRecorder().get_status(),
-        'player_status': player_status,
-        'timeline': timeline,
-    }
-
-
-def _transport_signature( recorder_status, player_status ):
-    """A token of the *button-relevant* state, so the client only fragment-
-    refreshes the transport when the available controls actually change (not
-    on every playhead tick or step-count bump)."""
-    return '{}|{}|{}|{}'.format(
-        int( recorder_status[ 'recording' ] ),
-        int( recorder_status[ 'paused' ] ),
-        player_status[ 'mode' ],
-        int( recorder_status[ 'has_working' ] ),
-    )
 
 
 class ScenesIndexView( View ):
@@ -341,16 +70,16 @@ class ScenesIndexView( View ):
             'timeline': None,
         }
         if scene is not None:
-            simulator_by_module = _simulator_by_module()
-            context[ 'module_rows' ] = self._build_module_rows( scene, simulator_by_module )
+            simulator_by_module_map = simulator_by_module()
+            context[ 'module_rows' ] = self._build_module_rows( scene, simulator_by_module_map )
             for control in scene.controls.all():
-                resolved = _resolve_control( control, simulator_by_module )
+                resolved = resolve_control( control, simulator_by_module_map )
                 if resolved is not None:
                     context[ 'control_grid' ].append( resolved )
                 else:
                     context[ 'missing_control_count' ] += 1
                 continue
-            context.update( _build_transport_context( scene, request.GET.get( 'sequence' )))
+            context.update( build_transport_context( scene, request.GET.get( 'sequence' )))
         return render( request, 'scenes/pages/scenes.html', context )
 
     def _resolve_selected_scene( self, request, scene_list ):
@@ -365,13 +94,13 @@ class ScenesIndexView( View ):
             return active
         return scene_list[0] if scene_list else None
 
-    def _build_module_rows( self, scene, simulator_by_module ):
+    def _build_module_rows( self, scene, simulator_by_module_map ):
         binding_by_module = {
             binding.module_key: binding
             for binding in scene.bindings.select_related( 'profile' ).all()
         }
         rows = []
-        for module_key, simulator in simulator_by_module.items():
+        for module_key, simulator in simulator_by_module_map.items():
             binding = binding_by_module.get( module_key )
             rows.append({
                 'module_key': module_key,
@@ -516,7 +245,7 @@ class SceneEditStatesView( View ):
             scene.controls.all().delete()
             controls = []
             for order_id, key in enumerate( keys ):
-                module_key, entity_name, sim_state_id = _split_control_key( key )
+                module_key, entity_name, sim_state_id = split_control_key( key )
                 controls.append( SimSceneControl(
                     scene = scene,
                     module_key = module_key,
@@ -530,7 +259,7 @@ class SceneEditStatesView( View ):
 
     def _build_state_tree( self, scene ):
         selected_keys = {
-            _control_key( c.module_key, c.entity_name, c.sim_state_id )
+            control_key( c.module_key, c.entity_name, c.sim_state_id )
             for c in scene.controls.all()
         }
         tree = []
@@ -541,7 +270,7 @@ class SceneEditStatesView( View ):
             for sim_entity in simulator.sim_entities:
                 states = []
                 for sim_state in sim_entity.sim_state_list:
-                    key = _control_key(
+                    key = control_key(
                         simulator.module_key, sim_entity.name, sim_state.sim_state_id,
                     )
                     dom_counter += 1
@@ -683,7 +412,7 @@ class SceneSequenceEditView( View ):
         sequence = _get_sequence( scene_id, sequence_id )
         name = ( request.POST.get( 'name' ) or '' ).strip()
         raw = request.POST.get( 'steps' ) or ''
-        steps, error = _parse_steps( raw )
+        steps, error = parse_steps( raw )
         if not error and not name:
             error = 'Name is required.'
         if not error and SimStateSequence.objects.filter(
@@ -712,7 +441,7 @@ class SceneSequenceExportView( View ):
             content_type = 'application/json',
         )
         response[ 'Content-Disposition' ] = (
-            f'attachment; filename="{_safe_filename( sequence.name )}.json"'
+            f'attachment; filename="{safe_filename( sequence.name )}.json"'
         )
         return response
 
@@ -733,7 +462,7 @@ class SceneSequenceImportView( View ):
             raw = upload.read().decode( 'utf-8', errors = 'replace' )
         else:
             raw = request.POST.get( 'steps' ) or ''
-        steps, error = _parse_steps( raw )
+        steps, error = parse_steps( raw )
         name = ( request.POST.get( 'name' ) or '' ).strip() or self._name_from_payload( raw )
         if not name:
             name = f'Sequence {scene.state_sequences.count() + 1}'
@@ -772,7 +501,7 @@ class SceneStatusView( View ):
         recorder_status = SimRecorder().get_status()
         player_status = SimPlayer().get_status()
         return JsonResponse({
-            'signature': _transport_signature( recorder_status, player_status ),
+            'signature': transport_signature( recorder_status, player_status ),
             'player': {
                 'mode': player_status[ 'mode' ],
                 'playing': player_status[ 'playing' ],
@@ -797,5 +526,5 @@ class SceneTransportView( View ):
 
     def get( self, request, scene_id, *args, **kwargs ):
         scene = _get_scene( scene_id )
-        context = _build_transport_context( scene, request.GET.get( 'sequence' ))
+        context = build_transport_context( scene, request.GET.get( 'sequence' ))
         return render( request, 'scenes/panes/transport.html', context )
